feat: Use floats to position legend and inputs
Initially, fieldsets used a flexbox solution involving rows, but it
didn't work due to a bug (https://stackoverflow.com/a/28078860). The
nested legend element inside fieldset didn't register as a flex-item,
but wrapping around a div seemed to fix it.

Only problem is accessibility since legends must be a direct descendant
of fieldsets for screen readers to work. Links:

https://www.accessibility-developer-guide.com/examples/forms/grouping-with-fieldset-legend/#legend-must-be-a-direct-child-of-fieldset
Solutions to issue (https://stackoverflow.com/a/59425373).

As a result, floats provide a decent alternative to flexbox in this
case.