refactor: Lift state up to FormFieldset
The main purpose in lifting state up from NewInputs to FormFieldset
is to readjust the identifier upon removing a <NewInputs />.

We incorrectly stored <NewInputs /> in FormFieldset state. Instead
store the data associated with the components and map it in render().
Fix this issue by storing data in this.state.info and not JSX.

To fix the identifier, pass a new prop called index (line 75) to <NewInputs />
in FormFieldset. This then gets passed all the way down to <Input /> for use.