refactor: Use controlled components in App
Previously, input values weren't stored in App state and was passed
as data in Form.jsx. At the time it was fine; the data was still being
passed to handleInputChange and processed to make a pdf. Only problem
was the addition of a new feature: persisting state through localStorage.

This soon proved to problematic as Form's data, this.formSection,
never passed a value prop to <Input />, which in turn alway left
the value attribute as undefined. Upon addition of a value string to
this.formSection, the input empties on every rerender, causing it to only
take a single character.

To solve this and have localStorage capabilities, include the value in
state and pass it to children as props.